refactor(calculator): CalculatorInput 컴포넌트 쪼개기
