+++ app/src/Calculator/constants/projectListTitles.ts
@@ -1,0 +1,8 @@
+export const PROJECT_LIST_TITLES = {
+  PROJECT_NAME: '프로젝트명',
+  SCORE: '점수',
+  COALITION_BONUS: '코알리숑 보너스',
+  EXP: '경험치',
+  FINISH_LEVEL: '통과 레벨',
+  BLACKHOLE: '블랙홀',
+} as const;
